+++ flang/lib/Optimizer/Dialect/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_subdirectory(Support)
 add_subdirectory(CUF)
+add_subdirectory(FIRCG)
 
 add_flang_library(FIRDialect
   FIRAttr.cpp
