+++ flang/lib/Optimizer/Dialect/FIRCG/CMakeLists.txt
@@ -1,0 +1,14 @@
+add_flang_library(FIRCodeGenDialect
+  CGOps.cpp
+
+  DEPENDS
+  CGOpsIncGen
+
+  LINK_LIBS
+  FIRDialect
+  MLIRIR
+
+  LINK_COMPONENTS
+  AsmParser
+  AsmPrinter
+  )
