+++ flang/lib/Optimizer/OpenACC/CMakeLists.txt
@@ -6,7 +6,6 @@ add_flang_library(FIROpenACCSupport
 
   DEPENDS
   FIRBuilder
-  FIRCodeGen
   FIRDialect
   FIRDialectSupport
   FIRSupport
@@ -14,7 +13,7 @@ add_flang_library(FIROpenACCSupport
 
   LINK_LIBS
   FIRBuilder
-  FIRCodeGen
+  FIRCodeGenDialect
   FIRDialect
   FIRDialectSupport
   FIRSupport
